Task 3 done

+++ src/main/java/at/ac/fhcampuswien/App.java
@@ -28,11 +28,11 @@ public class App {
      int a = 0xface;
      int b = 'Z';
      int c = 012;
-     long d = 80L;
-     float e = 44e-1f;
-     float f = 5.5f;
-     double g = 8.88e1;
-     double h = 99.9;
+     int d = 80;
+     int e = (int) 44e-1f;
+     int f = (int) 5.5f;
+     int g = (int) 8.88e1;
+     int h = (int) 99.9;
      int sum = (int) (a + b + c + d + e + f + g + h);
         System.out.println(sum);
